requirements/development.txt
# Install base requirements first
-r base.txt

# Development and debugging tools
django-debug-toolbar>=4.4.0
django-silk>=5.0.4

# Internationalization management (development only)
django-rosetta>=0.10.0

# Load testing
locust>=2.20.0

# Code quality
ipython>=8.18.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
